Login crash - Fixed

+++ Qoot/Qoot/Model/User/User.swift
@@ -38,10 +38,10 @@ class User: NSManagedObject {
         user.mobileNumber = userData.mobile_number
         user.name = userData.userName
         user.password = userData.password
-        user.registerID = Int16(userData.registration_id)
+        user.registerID = userData.registration_id
         user.registerType = userData.registration_type
         //user.urlStatus = userData.url_status
-        user.userId = Int16(userData.userId)
+        user.userId = userData.userId
         CoreDataHandler.sharedInstance.saveContext()
     }
